requirements.txt
python-dotenv
streamlit
pandas
streamlit-aggrid
langchain
langchain-community
chromadb
openai
